new templatetag ga_tag userVar. Make redirect_to_page decorator multi-app capable
