+++ facebook/views.py
@@ -1,13 +1,12 @@
-import functools, sys, logging, urllib2
+import sys, logging, urllib2
 from datetime import datetime, timedelta
 
 from django.conf import settings
-from django.http import HttpResponse, HttpResponseBadRequest, HttpResponseRedirect, HttpResponseForbidden
+from django.http import HttpResponse, HttpResponseBadRequest, HttpResponseForbidden,\
+    Http404
 from django.views.decorators.csrf import csrf_exempt
-from django.shortcuts import render_to_response, redirect, get_object_or_404, render
-from django.template.defaultfilters import urlencode
+from django.shortcuts import render_to_response, get_object_or_404, render
 from django.template import loader, RequestContext
-from django.core.urlresolvers import resolve, Resolver404, reverse
 
 from facebook.utils import validate_redirect
 from facebook.graph import get_graph
@@ -64,58 +63,6 @@ def input(request, action):
     return HttpResponseBadRequest('action %s not implemented' % action)
 
 
-def redirect_to_page(view):
-    """ Decorator that redirects a canvas URL to a page using the path that is in app_data.path """
-    """ Decorate the views where you have links to the app page. """
-
-    @functools.wraps(view)
-    def wrapper(*args, **kwargs):
-        request = args[0]
-        # if this is already the callback, do not wrap.
-        if getattr(request, 'avoid_redirect', False):
-            logger.debug('entered calback. View: %s, kwargs: %s' %(view, kwargs))
-            return view(*args, **kwargs)
-
-        session = request.session.get('facebook', dict())
-        try:
-            signed_request = session['signed_request']
-        except KeyError:
-            logger.debug('No signed_request in current session. Returning View.')
-            return view(*args, **kwargs)
-
-        logger.debug('signed_request: %s' %signed_request)
-
-        if 'app_data' in signed_request:
-            app_data = signed_request['app_data']
-            del request.session['facebook']['signed_request']['app_data']
-            request.session.modified = True
-            logger.debug('found app_data url: %s' %app_data)
-            #return HttpResponseRedirect(app_data)
-            try:
-                original_view = resolve(app_data)
-            except Resolver404:
-                logger.debug('Did not find view for %s.' %app_data)
-                url = u'%s?sk=app_%s' % (redirect_url, app_id)
-                return render_to_response('redirecter.html', {'destination': url }, RequestContext(request))
-
-            logger.debug('found original view url: %s' %original_view)
-            setattr(request, 'avoid_redirect' ,  True)
-            # call the view that was originally requested:
-            return original_view.func(request, *original_view.args, **original_view.kwargs)
-        else:
-            #check if the app is inside the specified page.
-            try:
-                page = signed_request['page']['id']
-            except KeyError:
-                page = None
-            if page <> page_id and not runserver:
-                logger.debug('Tab is not in original Page. Redirecting...')
-                url = u'%s?sk=app_%s&app_data=%s' % (redirect_url, app_id, urlencode(request.path))
-                return render_to_response('redirecter.html', {'destination': url }, RequestContext(request))
-
-        return view(*args, **kwargs)
-
-    return wrapper
 
 @csrf_exempt
 def channel(request):
